start-redis-reads/requirements.txt
functions-framework==3.*
google-cloud-storage
fsspec
gcsfs